Move build info loading to new BuildInfoManager managed object
Move build info loading to new BuildInfoManager managed object
