+++ src/main/groovy/edu/oregonstate/mist/api/InfoResource.groovy
@@ -14,19 +14,8 @@ import javax.ws.rs.core.Response
 class InfoResource extends Resource {
     private Info info = new Info()
 
-    public InfoResource() {
-        def stream = this.class.getResourceAsStream('/build.properties')
-        if (stream == null) {
-            throw new Exception("couldn't load build.properties")
-        }
-
-        def properties = new Properties()
-        properties.load(stream)
-
-        info.name = properties.get('name')
-        info.time = Long.parseLong(properties.getProperty('time'))
-        info.commit = properties.get('commit')
-        info.documentation = properties.get('documentation')
+    public InfoResource(Info info) {
+        this.info = info
     }
 
     /**
